+++ packages/blue-sdk-viem/src/index.ts
@@ -1,7 +1,9 @@
 export * from "./abis";
 export * from "./fetch";
 export * from "./types";
+export * from "./utils";
 
 export * as abis from "./abis";
 export * as fetch from "./fetch";
 export * as types from "./types";
+export * as utils from "./utils";