+++ packages/blue-sdk-viem/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 1.10.0 (2024-09-20)
+
+**Note:** Version bump only for package @morpho-org/blue-sdk-viem
+
+## 1.9.1 (2024-09-20)
+
+**Note:** Version bump only for package @morpho-org/blue-sdk-viem
+
+# 1.9.0 (2024-09-20)
+
+**Note:** Version bump only for package @morpho-org/blue-sdk-viem
+
+# 1.8.0 (2024-09-19)
+
+**Note:** Version bump only for package @morpho-org/blue-sdk-viem
+
 ## 1.7.5 (2024-09-16)
 
 **Note:** Version bump only for package @morpho-org/blue-sdk-viem